+++ Modules/Microsoft365DSC/Modules/M365DSCUtil.psm1
@@ -2589,7 +2589,11 @@ function Assert-M365DSCBlueprint
 
         [Parameter()]
         [System.String[]]
-        $ExcludedProperties
+        $ExcludedProperties,
+
+        [Parameter()]
+        [System.String[]]
+        $ExcludedResources
     )
 
     $InformationPreference = 'SilentlyContinue'
@@ -2636,21 +2640,40 @@ function Assert-M365DSCBlueprint
             $endPosition = $fileContent.IndexOf("`r", $startPosition)
             $fileContent = $fileContent.Remove($startPosition, $endPosition - $startPosition)
         }
-        $parsedBluePrint = ConvertTo-DSCObject -Content $fileContent
+
+        try
+        {
+            $parsedBluePrint = ConvertTo-DSCObject -Content $fileContent
+        }
+        catch
+        {
+            throw $_
+        }
 
         # Generate an Array of Resource Types contained in the BluePrint
         $ResourcesInBluePrint = @()
         foreach ($resource in $parsedBluePrint)
         {
+            if ($resource.ResourceName -in $ExcludedResources)
+            {
+                continue
+            }
             if ($ResourcesInBluePrint -notcontains $resource.ResourceName)
             {
                 $ResourcesInBluePrint += $resource.ResourceName
             }
         }
 
-        if (!$ResourcesInBluePrint)
+        if ([String]::IsNullOrEmpty($ResourcesInBluePrint))
         {
-            Write-Host 'Malformed BluePrint, aborting'
+            if (![String]::IsNullOrEmpty($ExcludedResources))
+            {
+                Write-Host 'All resources were excluded from BluePrint, aborting'
+            }
+            else
+            {
+                Write-Host 'Malformed BluePrint, aborting'
+            }
             break
         }
 
@@ -2681,7 +2704,8 @@ function Assert-M365DSCBlueprint
             -IsBlueprintAssessment:$true `
             -HeaderFilePath $HeaderFilePath `
             -Type $Type `
-            -ExcludedProperties $ExcludedProperties
+            -ExcludedProperties $ExcludedProperties `
+            -ExcludedResources $ExcludedResources
     }
     else
     {